Update patch set 3
Patch Set 3:

> So, should I bring back those Conflicts entries finally?

For me, with or without it are ok, however note that if you use the conflict you need to make sure you create a tag for new major release whenever is created in neutron or it will fail to build, breaking the consistency of the repo. For neutron-lbaas, fwaas and vpnaas it's ok because they are governed by the same team. Not sure about neutron-dynamic-routing, so i'd be in favor of not using it if possible.

Patch-set: 3