Update patch set 5
Patch Set 5: Code-Review-1

(1 comment)

We are doing something strange here. We define an openstack-neutron_dynamic_routing package that:

- requires the -common subpackage
- takes care of the systemd stuff in post/preun/postun

However, there are is no %files section for it, so there is no rpm created for it (see [1]).

We also have the openstack-neutron-bgp-dragent package, with the systemd unit file, but no package depends on it and it doesn't take care of the systemd post/preun/postun.

I guess we should merge the two packages into one?

[1] https://logs.rdoproject.org/35/7935/5/check/DLRN-rpmbuild/Z3a0953b7629c4992916ab43c2923e612/centos/current/

Patch-set: 5
Label: Code-Review=-1