Fix upgrade not updating common package
When upgrading the bgp-dragent package the
common package was not updated which results
in the python2 package not being updated and
the service not working.

This makes sure the agent package always depends
on the same release of the common package. The
common package then has a dependency for the python2
package so both of them should be upgraded correctly.

Change-Id: I685b7e53d40091652704593718524027e55faf94